+++ web/app/application.js
@@ -6,6 +6,11 @@ import './styles/main.scss';
 import '../../src/styles/prism.css';
 import '../../src/javascript/prism.js';
 
+//NOTE: need this to be an include other wise the configuration is not available in the constructor of components
+//TODO: figure out why this need to be an include and document it
+//set configuration values
+import './configuration-setup';
+
 //application bootstrap
 import './router';